Evaluation_framework_TindREd/evaluation/requirements.txt
# Framework de Evaluación - Mia Gente
# Instalar con: pip install -r requirements.txt

# Core
ragas>=0.1.0
langsmith>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-anthropic>=0.1.0

# Métricas y evaluación
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Reportes
jinja2>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0
aiohttp>=3.8.0

# JSON/YAML
pyyaml>=6.0
orjson>=3.9.0
